requirements.txt
openai
langgraph
python-dotenv
python-docx
streamlit
tiktoken
